*** empty log message ***
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@98743 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: ed14d7f200f56dd99217ede72e394205b4e04580

+++ Online/Online/Gaucho/src/components/UpdateAndReset.cpp
@@ -115,9 +115,6 @@ StatusCode UpdateAndReset::initialize() {
     taskName = serviceParts[2];
   }
   
-  
-  
-
   m_infoFileStatus = "SAVESETLOCATION/......................................................";
   std::string infoName = partName+"/"+taskName+"/SAVESETLOCATION";
   m_dimSvcSaveSetLoc = new DimService(infoName.c_str(),(char*)m_infoFileStatus.c_str());
@@ -345,7 +342,7 @@ ulonglong UpdateAndReset::gpsTime() {
   // this is only for test when Odin doesn't work
   ulonglong currentTime = GauchoTimer::currentTime();
 
-  int cycleNumber = currentTime/(m_desiredDeltaTCycle*1000000);
+  int cycleNumber = (int)currentTime/(m_desiredDeltaTCycle*1000000);
 
   ulonglong gpsTime = ((ulonglong)cycleNumber)*((ulonglong) m_desiredDeltaTCycle*1000000);
